Update install.rst
I added an installation guide that uses Docker to compile on a Raspberry Pi running Alpine Linux. Since the Alpine package manager is slightly different, I had te figure out which packages were missing in a default installation (namely, bash, make, and patch) and how some of the libraries have different names. I decided to share my (generic) dockerfile with the public, as it may be of help for others (and future me).