Disable some tests that HiGHS can not pass.
Some tests can't be used because HiGHS doesn't support some required feature, eg
cut generation, lazy constraints or MIP start. This affects many of the
examples (so we exclude all now) and also the tests "rcpsp" and "two_dim_pack".

Some others fail because the current version of HiGHS doesn't allow querying the
variable type. So we have to keep track of that in the wrapper, which only works
when the model is built in Python, but not when read from a file, ie in the
tests "mip_files".

Finally, some examples yielded a parse error, eg "1443_0-9.lp".

Was: Revert "include HiGHS as solvers in other (slow) tests". This (partially)
reverts commit 8918708369d0f4c61a5767eb25bfa8c2ee53c3fa.